pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "givondo-growth-engine"
version = "0.1.0"
description = "An autonomous multi-agent system that discovers opportunities and generates applications using Google ADK"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Givondo", email = "givondo@example.com" }
]
keywords = [
    "ai",
    "agents",
    "google-adk",
    "opportunity-discovery",
    "automation",
    "multi-agent",
    "llm"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # Google ADK - Core agentic framework
    "google-adk>=0.1.0",
    
    # Multi-model support
    "litellm>=1.40.0",
    "anthropic>=0.34.0",
    "openai>=1.40.0",
    "google-generativeai>=0.8.0",
    
    # Data layer
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "pinecone-client>=3.2.0",
    
    # Data validation & settings
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    
    # Async & HTTP
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    
    # Web scraping
    "playwright>=1.45.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "rich>=13.7.0",
    "tenacity>=8.5.0",
    "typer>=0.12.0",
    
    # Scheduling
    "apscheduler>=3.10.0",
    
    # API Server
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]
ml = [
    # Machine Learning
    "xgboost>=2.0.0",
    "scikit-learn>=1.5.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    
    # Intelligence Layer
    "google-generativeai>=0.8.0",
    "pinecone-client>=3.2.0",
]

[project.scripts]
growth-engine = "src.cli:main"
daily-run = "scripts.daily_run:main"

[project.urls]
Homepage = "https://github.com/givondo/growth-engine"
Documentation = "https://github.com/givondo/growth-engine#readme"
Repository = "https://github.com/givondo/growth-engine"
Issues = "https://github.com/givondo/growth-engine/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow",
    "e2e: marks tests as end-to-end (requires API keys)",
]
